backend_django/requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.4.0
django-countries==6.1.3
djangorestframework==3.11.1
mysqlclient==2.0.1
Pillow==7.2.0
PyMySQL==0.10.0
pytz==2020.1
sqlparse==0.3.1
